requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.4
django-ratelimit==4.1.0
gunicorn==25.3.0
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
whitenoise==6.12.0
